amsaaexactfail {SMRD}	R Documentation

AMSAA reliability growth test

Description

Bogdanoff and Kozin give the size of fatigue cracks as a function of number of the cycles and the of applied stress for 21 test specimens. This is an example of degradation data. The data were reported originally in Hudak, Saxena, Bucci, and Malcolm (1978). The data were collected to obtain information on crack growth rates for the alloy.

Format

A data.frame with 725 rows and 3 variables:

[, 1] 	 vehicle 	 Vehicle type 	 Categoric

[, 2] 	 miles 	 Accumulated distance at event 	 Numeric

[, 3] 	 event 	 Event observed at miles (Start/Fail/End) 	 Categoric

Source

Lieblein J., and Zelen, M. (1956), Statistical investigation of the fatigue life of deep-groove ball bearings, Journal of Research, National Bureau of Standards, 57, 273–316.

[Package SMRD version 0.13.1 Index]