SMRD {SMRD}	R Documentation

SMRD: Comprehensive Life-Data Analysis in R

Description

SMRD: Comprehensive Life-Data Analysis in R

Details

Since its publication, the text Statistical Methods for Reliability Data has been recognized as a foundational resource in analyzing time to event data. Along with the text, the authors also developed an S-Plus software pacakge, called SPLIDA, to help readers utilize the methods presented in the text. Today, R is is the most popular statistical computing language in the world, largely supplanting S-Plus. The SMRD package is the result of a multi-year effort to completely rebuild SPLIDA, to take advantage of the improved graphics and workflow capabilities available in R. This presentation introduces the SMRD package, outlines the improvements and shows how the package works seamlessly with the rmarkdown and shiny packages to dramatically speed up your workflow. The presentation concludes with a discussion on what improvements still need to be made prior to publishing the package on the CRAN

[Package SMRD version 0.13.1 Index]