adhesivestrength {SMRD}	R Documentation

Accelerated adhesive creep test

Description

Data from fatigue endurance tests. The ball bearings came from four different major bearing companies. There was disagreement in the industry on the appropriate parameter values to use to describe the relationship between fatigue life and stress loading. The main objective of the study was to estimate values of the parameters in the equation relating bearing life to load.

Format

A data.frame with 89 rows and 4 variables:

[, 1] 	 pounds 	 Load at which the adhesive failed 	 Numeric

[, 2] 	 celsius 	 Temperature applied to the bond 	 Categoric

[, 3] 	 days 	 Time the bond was under load 	 Categoric

[, 4] 	 group 	 Type of adhesive	 Categoric

Source

Beckwith J.P. (1980), An estimator and design techinque for the estimation of a rate parameter in accelerated testing. Department of Mathematics and Computer Science, Michigan Technological University, Houghton, MI.

[Package SMRD version 0.13.1 Index]