alloya {SMRD}	R Documentation

Fatigue crack-growth test (Alloy-A)

Description

Dataset: Crack lengths observed after each increment of 10,000 fatigue cycles.

Format

A data.frame with 262 rows and 3 variables

[, 1] 	 inches 	 Crack length observed at megacycles 	 Numeric

[, 2] 	 specimen 	 Test specimen designator 	 Categoric

[, 3] 	 megacycles 	 Number of fatigue cycles (in millions) when length was inspected 	 Numeric

Details

Twenty-one specimens of an unknown alloy were subjected to the same cyclic load profile. Prior to testing, an initial crack (length = 0.9 in) was cut into each specimen to serve as a crack nucleation site. After each increment of 10,000 fatigue cycles the length of the crack was recorded. A specimen was considered to have failed when the crack reached 1.6 inches. The test concluded when the crack length reached 1.6 inches or after 1.2 million cycles.

Note

These data were obtained from Figure 4.5.2 of Bogdanoff and Kozin (1985, pg 242)

Source

S.J. Hudak, A. Saxena, R.J. Bucci, and R.C. Malcom (1978), Development of standard methods of testing and analyzing fatigure crack growth rate data, Technical Report AFML-TR-78-40, Westinghouse R & D Center, Westinghouse Electric Corporation, Pittsburgh, PA 15235.

J.L. Bogdanoff and F. Kozin, (1985), Probablistic Models of Cumulative Damage, John Wiley & Sons, New York

[Package SMRD version 0.13.1 Index]