alloyc {SMRD}	R Documentation

Tensile strength test (Alloy-C)

Description

These data resulted from a test that was conducted to obtain information on the strength of an alloy produced by an alternate process. Failure times were recorded in intervals of strength, measure in ksi (1 ksi = 1000 psi)

Format

A data.frame with 9 rows and 4 variables:

[, 1] 	 ksil 	 Start of an observation interval	 Numeric

[, 2] 	 ksiu 	 End of an observation interval	 Numeric

[, 3] 	 event 	 Event observed in the interval (interval-censored)	 Categoric

[, 4] 	 count 	 Number of events observed in the interval	 Numeric

Source

Meeker W.Q. and Escobar L.A., Statistical Methods for Reliability Data, New York: John Wiley & Sons. 276

[Package SMRD version 0.13.1 Index]