appliancea {SMRD}	R Documentation

Appliance cord failures

Description

Bogdanoff and Kozin give the size of fatigue cracks as a function of number of the cycles and the of applied stress for 21 test specimens. This is an example of degradation data. The data were reported originally in Hudak, Saxena, Bucci, and Malcolm (1978). The data were collected to obtain information on crack growth rates for the alloy.

Format

A data.frame with 36 rows and 3 variables:

[, 1] 	 hours 	 Accumulated time at event 	 Numeric

[, 2] 	 event 	 Event observed at hours (failed/right-censored) 	 Categoric

[, 3] 	 type 	 Appliance type 	 Categoric

Source

Nelson W. (1982), Applied Life Data Analysis, New York: John Wiley & Sons, pg 121

[Package SMRD version 0.13.1 Index]