applianceb {SMRD}	R Documentation

Appliance-B failures (multiple data sources)

Description

Bogdanoff and Kozin give the size of fatigue cracks as a function of number of the cycles and the of applied stress for 21 test specimens. This is an example of degradation data. The data were reported originally in Hudak, Saxena, Bucci, and Malcolm (1978). The data were collected to obtain information on crack growth rates for the alloy.

Format

A data.frame with 398 rows and 5 variables:

[, 1] 	 source 	 Information source (test environment) 	 Categoric

[, 2] 	 event 	 Event observed at days (failure/right-censored/left-censored/interval-censored) 	 Categoric

[, 3] 	 count 	 Number of events observed at time 	 Numeric

[, 4] 	 days 	 Accumulated time at event 	 Numeric

[, 5] 	 mode 	 Failure mode observed at days 	 Categoric

Source

Lieblein J., and Zelen, M. (1956), Statistical investigation of the fatigue life of deep-groove ball bearings, Journal of Research, National Bureau of Standards, 57, 273–316.

[Package SMRD version 0.13.1 Index]