aprel72 {SMRD}	R Documentation

#####

Description

Bogdanoff and Kozin give the size of fatigue cracks as a function of number of the cycles and the of applied stress for 21 test specimens. This is an example of degradation data. The data were reported originally in Hudak, Saxena, Bucci, and Malcolm (1978). The data were collected to obtain information on crack growth rates for the alloy.

Format

A data.frame with 25 rows and 5 variables:

[, 1] 	 hoursl 	 Start of observation interval 	 Numeric

[, 2] 	 hoursu 	 End of observation interval 	 Numeric

[, 3] 	 event 	 Event observed in the interval (right-censored/left-censored/interval-censored) 	 Categoric

[, 4] 	 count 	 Number of events observed in the interval 	 Numeric

[, 5] 	 celsius 	 Temperature applied 	 Categoric

Source

Meeker W.Q. and Escobar L.A. (1998), Statistical Methods for Reliability Data

[Package SMRD version 0.13.1 Index]