at7987 {SMRD}	R Documentation

Fatigue life test (alloy T7987)

Description

A test was conducted to assess the fatigue life of an alloy, designated as T-7987.

Format

A data.frame with 68 rows and 3 variables:

[, 1] 	 kilocycles 	 Accumulated cycles at event (in thousands) 	 Numeric

[, 2] 	 event 	 Event observed at kilocycles (failed/right-censored) 	 Categoric

[, 3] 	 count 	 Number of events observed at kilocycles 	 Numeric

Details

In this test, 72 specimen were subjected to an unreported load spectrum. The number of cycles at which a specimen failed was recorded (rounded to the nearest thousand) up to 300 kilocycles. Of the 72 units tested, 67 failures were observed and 5 units were right censored at 300 kilocycles.

Source

Meeker W.Q. and Escobar L.A. (1998) Statistical Methods for Reliability Data, New York: John Wiley & Sons.

[Package SMRD version 0.13.1 Index]