bearingcage {SMRD}	R Documentation

Large system bearing cage test

Description

Ball bearing assembly failure data

Format

A data.frame with 25 rows and 3 variables:

[, 1] 	 hours 	 Accumulated time at event 	 Numeric

[, 2] 	 event 	 Event observed at hours (failure/right-censored/left-censored/interval-censored) 	 Categoric

[, 3] 	 count 	 Number of events observed at hours 	 Numeric

Details

The service life requirement for a ball bearing assembly was specified such that t[0.1] (aka the B10 life) be greater that 8000 hours. Analysts were concerned that the design of the bearing cage in this assembly was inadequate and could lead to premature failures during service. Service times were collected for 1703 assemblies that were introduced into service over time. The analysts wanted to use the service life data to determine if a redesign was needed to ensure that the units could meet the service life requirement. Management was also interested in determining the number of additional failures that could be expected over the next year for the population of assemblies already in service.

Note

Bearing cages are used in ball bearing assemblies to ensure that the ball bearings do not drift out of position relative to the other ball bearings during use.

Source

R. B. Abernethy, J. E. Breneman, C. H. Medlin, and G. L. Reinman, (1983) Weibull Analysis Handbook, Air Force Wright Aeronautical Laboratories Technical Report AFWAL-TR-83-2079

[Package SMRD version 0.13.1 Index]