boot.pivot {SMRD}	R Documentation

Bootstrap based prediction intervals

Description

Bootstrap based prediction intervals

Usage

boot.pivot(B = 100, N = 20, mu = 5.13, sigma = 0.161, t_c = NULL,
  fails = NULL, alpha = 0.05)

Arguments

B
The number of bootstrap cycles

N
The size of the original data sample

mu
ML estimate for the mean of the original data sample

sigma
ML estimate for the standard deviation of the original data sample

t_c
Censoring time for Type-I censored tests

fails
Number of failures for Type-II censored tests

alpha
Desired significance level

[Package SMRD version 0.13.1 Index]