censored.data.plot {SMRD}	R Documentation

Plot failures and censored observations as a function of the response

Description

Plot failures and censored observations as a function of the response

Usage

censored.data.plot(data.ld, explan.var = 1, ylim = c(NA, NA), xlim = c(NA,
  NA), x.axis = "linear", y.axis = "linear", my.title = NULL,
  ylab = get.time.units(data.ld), xlab = NULL, cex = par()$cex,
  cex.labs = par()$cex.lab, cex.points = 1.2, add = F, grids = F,
  title.option = GetSMRDDefault("SMRD.TitleOption"), response.on.yaxis = T)

Arguments

data.ld
Name (as a character string) of a life.data object created using the function frame.to.ld

explan.var
numeric(1)

Explanatory variable to be plotted versus the responses. Value cannot be greater that the number x.columns defined in data.ld. Setting explan.var = 1 selects the x.column with the lowest number, regardless of the order in which the x.columns were defined in frame.to.ld. For example, data.ld can have argument x.columns = c(4,5,6) or x.columns = c(6,5,4). For either case, explan.var = 1 selects x.column = 4 since this is the smallest value. Note, explan.var may also be defined using a character string matching the name of column

ylim
numeric(c(NA, NA))

(optional) Range of plotting values for the y-axis. If left blank, values will be chosen based on the information in data.ld

xlim
numeric(c(NA, NA))

(optional) Range of plotting values for the x-axis. If left blank, values will be chosen based on the information in data.ld

x.axis
character("linear")

Transformation applied to the values defined along the x-axis. See SMRD:::.generic.relationship.name for the list of transformation relationships that can be used

y.axis
character("linear")

Transformation applied to the values defined along the x-axis. See SMRD:::.generic.relationship.name for the list of transformation relationships that can be used.

my.title
character(NULL)

Title to be displayed above the plot

ylab
character(get.time.units(data.ld))

(optional) Title of the y-axis. If left blank, the value of the time.units argument defined in data.ld will be used. If no time.units were defined when data.ld was created, 'Hours' is used as the default

xlab
character(NULL)

(optional) Title of the x-axis. If left blank, the column name in data.ld corresponding the argument explan.var is used.

cex
numeric(par()$cex)

cex.labs
numeric(par()$cex.lab)

cex.points
numeric(1.2)

Scaling value for size of the points to be plotted

add
logical(FALSE)

If TRUE plotting is suppressed to allow for the addition of more graphical features

grids
logical(FALSE)

If TRUE, gridlines are provided on the plot

title.option
character(GetSMRDDefault("SMRD.TitleOption"))

(optional) Produces a title above the plot using the SMRD default title format

response.on.yaxis
character(TRUE)

If FALSE, the x and y axes are reversed such that the response values are plotted along the x-axis

Author(s)

William Q. Meeker, PhD

See Also

frame.to.ld

Examples

# Create a \code{life.data} object
superalloy.ld <- frame.to.ld(superalloy,
                             response.column = 1, 
                             censor.column = 2,
                             x.columns = c(5,6,4),
                             time.units = "Kilocycles")

summary(superalloy.ld)

censored.data.plot(superalloy.ld, 
                  explan.var = 3)

censored.data.plot(superalloy.ld, 
                  explan.var = 3 ,
                  response.on.yaxis = F)

censored.data.plot(superalloy.ld, 
                  explan.var = 3, 
                  x.axis = "log", 
                  y.axis = "log")

censored.data.plot(superalloy.ld, 
                  explan.var = 3 ,
                  response.on.yaxis = F, 
                  x.axis = "log", 
                  y.axis = "log")

[Package SMRD version 0.13.1 Index]