cylinder {SMRD}	R Documentation

Diesel engine cylinder replacements

Description

Nelson and Doganaksoy (1989) present data on cylinder replacement times for 120 diesel engines. We take these engines to be a sample from a larger population of engines.Each engine has 16 cylinders.

Format

A data.frame with 276 rows and 4 variables:

[, 1] 	 engine 	 Type of engine used 	 Categoric

[, 2] 	 days 	 Accumulated time at event 	 Numeric

[, 3] 	 count 	 Number of events observed at days 	 Numeric

[, 4] 	 event 	 Event observed at days (Replacement/End) 	 Categoric

Details

Cylinders in a type of diesel engine can develop leaks or have low compression. Cylinders are inspected at times of convenience, along with other usual engine maintenance operations. Faulty cylinders are replaced by a rebuilt cylinder. More than one cylinder could be replaced at an inspection. Management needed to know if the company should perform preventive replacement of cylinders before they develop low compression.

Source

Nelson and Doganaksoy (1989)

[Package SMRD version 0.13.1 Index]