fan {SMRD}	R Documentation

Diesel generator fan failures

Description

These data are diesel generator fan failures. Failures in 12 of 70 generator fans were reported at times ranging between 450 hours and 8,750 hours. Of the 58 units that did not fail, the reported running times (i.e., censoring times) ranged between 460 and 11,500 hours. Different fans had different running times because units were introduced into service at different times and because their use-rates differed.

Format

A data.frame with 37 rows and 3 variables:

[, 1] 	 hours 	 Accumulated time at event 	 Numeric

[, 2] 	 event 	 Event observed at hours (failure/right-censored) 	 Categoric

[, 3] 	 count 	 Number of events observed at hours 	 Numeric

Source

Nelson, W. (1982), Applied Life Data Analysis, pg 133, New York, John Wiley & Sons Inc.

[Package SMRD version 0.13.1 Index]