frame.to.ld {SMRD}	R Documentation

Create a life.data object from a data.frame

Description

Create a life.data object from a data.frame

Usage

frame.to.ld(frame, response.column, censor.column, case.weight.column,
  failure.mode.column, truncation.type.column, truncation.response.column,
  time.units = names.the.frame[response.column[1]], x.columns = NULL,
  xlabel = NULL, data.title = deparse(substitute(frame)), data.note = "",
  func.call = match.call(), residual.rmd = NULL)

Arguments

frame
A data.frame class object.

response.column
The numeric index or name (as a character string) of the column(s) in frame containing the responses.

For reliability data, responses are often the amount of usage (measured in time, distance, cycles) that are recorded when an event occurred.

Events for which responses should be recorded include when one or more units fail or are censored. If the responses are recorded as an interval (i.e. an event occurs between a and b), response.column will accept vector arguments.

censor.column
The numeric index or name (as a character string) of the column in frame that contains the type(s) of censoring (exact failures, right-censoring, left-censoring, interval censoring) occurring at each event.

In practice, several different labels are used to refer to differnt types of censoring. frame.to.ld can accept many of these different labels. To view the list of see SMRDOptionsDefaults.

case.weight.column
The numeric index or name (as a character string) of the column in frame containing the number of events occurring at each entry in the response.column. (optional) This column may be left undefined, in which case it will be assumed that only one unit was observed to have failed/censored for each entry in response.column.

failure.mode.column
The numeric index or name (as a character string) of the column in frame containing the type(s) of failure modes observed. (optional) This column may be left undefined, in which case it will be assumed that only one failure mode was observed

truncation.type.column

truncation.response.column
The time at which an obsevation is trucated.

time.units
A character string denoting the unit of measure used to quantify system lifetime. time.units should be provided for each life.data object as this string will be automatically substituted in many SMRD plots.

x.columns

xlabel

data.note

func.call

residual.rmd

Author(s)

William Q. Meeker, PhD

See Also

frame.to.rmd, frame.to.ddd, surv.to.ld, SMRDOptionsDefaults

Examples

vignette('echapter1')
vignette('echapter3')
vignette('echapter6')

[Package SMRD version 0.13.1 Index]