gaaslaser {SMRD}	R Documentation

Gallium-Arsenic Laser Device Degradation Test Data

Description

Dataset: Observed increases in operating current after 250 hour intervals

Format

A data.frame with 255 rows and 3 variables:

[, 1] 	 increase 	 Percent increase in observed operating current 	 Numeric

[, 2] 	 unit 	 Unit designator 	 Categoric

[, 3] 	 hours 	 Accumulated time when increase was measured 	 Categoric

Details

Over the life of some lasers devices, degradation causes a decrease in light output. In this test, a feedback mechanism was used to maintain a constant light output by increasing the operating current. Fifteen GaAs laser devices were tested at an elevated temperature of 80^o C to accelerate the degradation process. The operating current was measured in 250-hour intervals and the data were recorded as the percent increase in current compared to the initial operating current measured when the test began. A device was considered to have failed if the percent increase reached 10%.

Source

Meeker W.Q. and Escobar L.A. (1998) Statistical Methods for Reliability Data, New York: John Wiley & Sons.

[Package SMRD version 0.13.1 Index]