gadget_ld {SMRD}	R Documentation

SMRD Analysis Shiny Gadget

Description

Shiny gadget used to visually inspect column values in a data set and subset rows by specifying column values

Usage

gadget_ld(data, theme = "flatly", width = "100%", height = "600px",
  css = NULL)

Arguments

data
A data set

theme
character A bootswatch theme provided to shinythemes::shinytheme

width
character Width of the gadget (in valid css units)

height
character Height of the gadget (in valid css units)

css
character Path to a custom css file

Value

A list of length 4

obj
A life.data object

dist
The selected distribution

cdfest
A cdfest object

mlest
A mlest object

Examples

## Not run: gadget_ld(data = shockabsorber)

[Package SMRD version 0.13.1 Index]