grids1 {SMRD}	R Documentation

Locomotive braking grid replacements

Description

Age of locomotives when either their braking grids were replaced or the largest age observed for each locomotive

Format

A data.frame with 39 rows and 3 variables:

[, 1] 	 unit 	 Unit type 	 Categoric

[, 2] 	 days 	 Accumulated time at event 	 Numeric

[, 3] 	 event 	 Event observed at days (Replacement/End) 	 Categoric

Details

Two batches grids1 & grids2

Source

Doganaksoy, N., and Nelson, W. (1998)

[Package SMRD version 0.13.1 Index]