hcfdata {SMRD}	R Documentation

#####

Description

Berkson investigated the randomness of alpha-particle emissions of americium-241 (which has a half-life of about 458 years). Physical theory suggests that, over a short period of time, the interarrival times of observed particles would be independent and come from an exponential distribution where the rate parameter is the mean time between arrivals. The corresponding homogeneous Poisson process that counts the number of emissions on the real-time line has arrival rate or the intensity lambda=1/theta. The data consist of 10,220 observed interarrival times of alpha particles (time unit equal to 1/5,000 second). The observed interarrival times were put into intervals (or bins) running from 0 to 4,000 time units with interval lengths ranging from 25 to 100 time units, with one additional interval for observed times exceeding 4,000 time units. To save space, this example uses a smaller number of larger bins; reducing the number of bins in this way will not seriously affect the precision of ML estimates.

Format

A data.frame with 83 rows and 9 variables:

[, 1] 	 specimen 	 Specimen type 	 Categoric

[, 2] 	 step 	 Was step-up loading used? 	 Categoric

[, 3] 	 hertz 	 Load frequency 	 Categoric

[, 4] 	 ratio 	 Stress ratio max(tensile stress)/max(compressive stress) 	 Categoric

[, 5] 	 cycles 	 Accumulated cycles at event 	 Numeric

[, 6] 	 range 	 Stress range max(tensile stress) - max(compressive stress) 	 Numeric

[, 7] 	 tensile 	 Maximum tensile stress 	 Numeric

[, 8] 	 event 	 Event observed at cycles (failed/right-censored) 	 Categoric

[, 9] 	 swt 	 Smith-Watson-Topper parameter 	 Numeric

Details

SWT = 0.5 x σΔε

Source

W.Q. Meeker, L.A. Escobar, and Lu (1998)

[Package SMRD version 0.13.1 Index]