heatexchanger {SMRD}	R Documentation

Nuclear power plant heat exchanger crack data

Description

These circuit packs were manufactured under the same design, but by two different vendors. The trial ran for 10,000 hours. The 4993 circuit packs from Vendor 1 came straight from production. The 4993 circuit packs from Vendor 2 had already seen 1000 hours of burn-in testing at the manufacturing plant under operating conditions similar to those in the field trial. The circuit packs manufactured by Vendor 2 were sold at a higher price because field reliability was supposed to have been improved by the burn-in screening of circuit packs containing defective components. Failures during the first 1000 hours of burn-in were not recorded. This is the reason for the unknown entries in the table and for having information out to 11,000 hours for Vendor 2. These data are for the first failure in a position; information on circuit packs replaced after initial failure in a position was not part of the study.

Format

A data.frame with 9 rows and 5 variables:

[, 1] 	 lower 	 Start of an inspection interval (in years) 	 Numeric

[, 2] 	 upper 	 End of an inspection interval (in years) 	 Numeric

[, 3] 	 event 	 Event observed in the interval (left-censored/right-censored/interval-censored)	 Categoric

[, 4] 	 count 	 Number of events observed in the interval 	 Numeric

[, 5] 	 plant 	 Plant where the tube was installed 	 Categoric

Source

Meeker, W. Q. and Escobar, L. A. (1996), Statistical Methods for Reliability Data, New York, Wiley-Interscience

[Package SMRD version 0.13.1 Index]