lfp1370 {SMRD}	R Documentation

Integrated circuit life test

Description

The primary purpose of this experiment was to estimate the proportion of defective units being manufactured in the current production process and to estimate the amount of 'burn-in' time that would be required to remove most of the defective units from the product population. The reliability engineers were also interested in whether it might be possible to get the needed information about the state of the production process, in the future, using much shorter tests (say 200 or 300 hours).\ The event column indicates that these data are singly right censored at 1370 hours. However, the presence of ties indicates that the data are inspection times, and are thus interval censored observations.

Format

A data.frame with 22 rows and 3 variables:

[, 1] 	 hours 	 Accumulated time at event 	 Numeric

[, 2] 	 event 	 Event observed at hours (failure/right-censored) 	 Categoric

[, 3] 	 count 	 Number of events observed at hours 	 Numeric

Source

W.Q. Meeker (1987), Limited failure population life tests: application to integrated circuit reliability, Technometrics, 29, 51-65.

See Also

lfptrun100

[Package SMRD version 0.13.1 Index]