lzbearing {SMRD}	R Documentation

Ball bearing fatigue test data

Description

The ball bearings came from four different major bearing companies. There was disagreement in the industry on the appropriate parameter values to use to describe the relationship between fatigue life and stress loading. The main objective of the study was to estimate values of the parameters in the equation relating bearing life to load.

Format

A data.frame with 23 rows and 1 variable:

[, 1] 	 megacycles 	 Accumulated cycles at failure (in millions) 	 Numeric

Source

J.F. Lawless (1982), Statistical Models and Methods for Lifetime Data, New York, Wiley & Sons

[Package SMRD version 0.13.1 Index]