machineh {SMRD}	R Documentation

Earth-moving machine maintenance

Description

The ball bearings came from four different major bearing companies. There was disagreement in the industry on the appropriate parameter values to use to describe the relationship between fatigue life and stress loading. The main objective of the study was to estimate values of the parameters in the equation relating bearing life to load.

Format

A data.frame with 573 rows and 4 variables:

[, 1] 	 unit 	 Unit type 	 Categoric

[, 2] 	 hours 	 Accumulated time at event 	 Numeric

[, 3] 	 cost 	 Price of test 	 Numeric

[, 4] 	 event 	 Event observed at hours (Action/End) 	 Categoric

Source

Meeker and Escobar (1998)

[Package SMRD version 0.13.1 Index]