piccioto {SMRD}	R Documentation

#####

Description

Picciotto includes the length and kcycles columns. picciotto2 is a subset of the picciotto data set, including only the units for which length = 30,60, or 90. picciotto3 mirrors picciotto2 but assumes that the test conlcuded at 100 kilocycles

Format

A data.frame with 797 rows and 3 variables:

[, 1] 	 inches 	 Length of the test specimen 	 Categoric

[, 2] 	 kilocycles 	 Accumulated cycles at event (in thousands) 	 Numeric

[, 3] 	 event 	 Event observed at kilocycles (failed/right-censored) 	 Categoric

Source

Picciotto R., (1970) Tensile Fatigue Characteristics of a Sized Polyester/Viscose Yarn and Their Effect on Weaving Performance. A thesis submitted to the Graduate Faculty of North Carolina State University at Raleigh in partial fulfillment of the requirements for the Degree of Master of Science. Department of Textile Technology.

[Package SMRD version 0.13.1 Index]