r4490 {SMRD}	R Documentation

Car door lock transmitter replacement data

Description

R4490 is the code used by General Motors denoting Remote Control Door Lock Transmitter Replacement

Format

A data.frame with 47 rows and 4 variables:

[, 1] 	 vin 	 Vehicle identifcation number 	 Categoric

[, 2] 	 days 	 Accumulated time at event 	 Numeric

[, 3] 	 event 	 Event observed at days (R4490/End/MEnd) 	 Categoric

[, 4] 	 costcount 	 What is cost count? 	 Numeric

Source

Shimokawa, T., and Hamaguchi, Y. (1987), Statistical Evaluation of Fatigue Life and Fatigue Strength in Circular-Holed Notched Specimens of a Carbon Eight-Harness-Satin/Epoxy Laminate,” in Statistical Research on Fatigue and Fracture (Current Japanese Materials Research, Vol. 2), eds. T. Tanaka, S. Nishijima, and M. Ichikawa, London: Elsevier, pp. 159-176.

[Package SMRD version 0.13.1 Index]