superalloy {SMRD}	R Documentation

#####

Description

Data from fatigue endurance tests. The ball bearings came from four different major bearing companies. There was disagreement in the industry on the appropriate parameter values to use to describe the relationship between fatigue life and stress loading. The main objective of the study was to estimate values of the parameters in the equation relating bearing life to load.

Format

A data.frame with 23 rows and 6 variables:

[, 1] 	 kilocycles 	 Accumulated cycles at event (in thousands) 	 Numeric

[, 2] 	 event 	 Event observed at kcycles (failure/right-censored) 	 Categoric

[, 3] 	 count 	 Number of events observed at kcycles 	 Numeric

[, 4] 	 pstress 	 Pseudo-stress applied derated 	 Numeric

[, 5] 	 lpstress 	 Log[pseudo-stress] (ln[pstress])The log base e transformation of the Pseudo-Stress 	 Numeric

[, 6] 	 lpstress2 	 Square of log[pseudo-stress] (ln[pstress]^2) 	 Numeric

Source

Nelson (1990)

[Package SMRD version 0.13.1 Index]