v7tube {SMRD}	R Documentation

V7 vacuum-tube life test data

Description

Berkson investigated the randomness of alpha-particle emissions of americium-241 (which has a half-life of about 458 years). Physical theory suggests that, over a short period of time, the interarrival times of observed particles would be independent and come from an exponential distribution where the rate parameter is the mean time between arrivals. The corresponding homogeneous Poisson process that counts the number of emissions on the real-time line has arrival rate or the intensity lambda=1/theta. The data consist of 10,220 observed interarrival times of alpha particles (time unit equal to 1/5,000 second). The observed interarrival times were put into intervals (or bins) running from 0 to 4,000 time units with interval lengths ranging from 25 to 100 time units, with one additional interval for observed times exceeding 4,000 time units. To save space, this example uses a smaller number of larger bins; reducing the number of bins in this way will not seriously affect the precision of ML estimates.

Format

A data.frame with 5 rows and 4 variables:

[, 1] 	 lower 	 Start of an inspection interval (in days) 	 Numeric

[, 2] 	 upper 	 End of an inspection interval (in days) 	 Numeric

[, 3] 	 event 	 Event observed in the interval (right-censored/left-censored/interval-censored) 	 Categoric

[, 4] 	 count 	 Number of events observed in the interval 	 Numeric

Source

Shimokawa, T., and Hamaguchi, Y. (1987), Statistical Evaluation of Fatigue Life and Fatigue Strength in Circular-Holed Notched Specimens of a Carbon Eight-Harness-Satin/Epoxy Laminate,” in Statistical Research on Fatigue and Fracture (Current Japanese Materials Research, Vol. 2), eds. T. Tanaka, S. Nishijima, and M. Ichikawa, London: Elsevier, pp. 159-176.

[Package SMRD version 0.13.1 Index]