zelencap {SMRD}	R Documentation

Glass capacitor accelerated life test data

Format

A data.frame with 40 rows and 5 variables:

[, 1] 	 hours 	 Accumulated time at event 	 Numeric

[, 2] 	 event 	 Event observed at hours (failure/right-censored) 	 Categoric

[, 3] 	 count 	 Number of events observed at hours 	 Numeric

[, 4] 	 celsius 	 Temperature applied 	 Categoric

[, 5] 	 volts 	 Electrical power applied 	 Categoric

Details

Eight capacitors were tested at each combination of temperature and voltage. For each combination the test was terminated after the fourth failure, yielding failure (Type II) censored data.

Source

Zelen (1959)

[Package SMRD version 0.13.1 Index]